* upstream/main:
  forgot about this one
  kollumos license
  HMMM
  initial sounds version
  adding snazzy purple man for next email campaign
  Remove jobs on fail
  more cache
  some fixes to avoid sending bsky posts to fedi
  make nsfw images load but with grayscale filter
  make media not lazyload. maybe bad idea?
  will this work?
  mail campaign for server update
  fixes
  Update README.md

+++ packages/backend/routes/deletePost.ts
@@ -34,7 +34,7 @@ const deletePostQueue = new Queue('deletePostQueue', {
       type: 'exponential',
       delay: 1000
     },
-    removeOnFail: 25000
+    removeOnFail: true
   }
 })
 
@@ -66,7 +66,7 @@ export default function deletePost(app: Application) {
           }
         }
         // bsky delete
-        if (postToDelete.bskyUri) {
+        if (postToDelete.bskyUri && user.enableBsky) {
           const agent = await getAtProtoSession(user)
           if (postToDelete.bskyCid) {
             await agent.deletePost(postToDelete.bskyUri)
